Add checks for php 8.1
